--- src/app/(main)/todo-list/_components/AddTodoModal.tsx
@@ -1,9 +1,0 @@
-// import Modal from "react-modal";
-
-// Modal.setAppElement("#__next");
-
-// const AddTodoModal = () => {
-//   return <Modal>모달</Modal>;
-// };
-
-// export default AddTodoModal;
