+++ src/todos/components/QuickAddTodoForm.tsx
@@ -1,9 +1,6 @@
 import { useState } from "react";
-import { useTodos } from "../useTodos";
-import { Todo } from "../types";
 import { AddTodoFormData } from "./AddTodoForm";
-import { Input } from "@/shared/ui/input";
-import { Button } from "@/shared/ui/button";
+import { IoIosAddCircleOutline } from "react-icons/io";
 
 export interface QuickAddTodoFormProps {
   onSubmit?: (data: AddTodoFormData) => void;
@@ -25,16 +22,22 @@ const QuickAddTodoForm = ({ onSubmit }: QuickAddTodoFormProps) => {
   };
 
   return (
-    <form onSubmit={handleSubmit}>
-      <Button type="submit">+</Button>
-      <Input
-        type="text"
-        placeholder="새로운 할 일을 입력하세요."
-        value={formTitle}
-        onChange={(e) => setFormTitle(e.target.value)}
-        autoFocus
-      />
-    </form>
+    <div>
+      <form
+        onSubmit={handleSubmit}
+        className="flex items-center border border-pai-100 border-solid rounded-[32px] bg-whiteTrans-wh72 p-4 mb-6"
+      >
+        <IoIosAddCircleOutline type="submit" className="w-[36px] h-[36px] text-gray-700" />
+        <input
+          type="text"
+          placeholder="투두리스트를 작성해보세요."
+          value={formTitle}
+          onChange={(e) => setFormTitle(e.target.value)}
+          autoFocus
+          className="outline-none p-2 w-[343px] h-[76p] placeholder-gray-400"
+        />
+      </form>
+    </div>
   );
 };
