+++ src/todos/components/TodoListPage.tsx
@@ -1,16 +1,14 @@
 "use client";
 
 import Calendar, { CalendarEvent } from "@/shared/ui/Calendar";
-import TodoList from "./TodoList";
-import { Todo } from "../types";
 import { useMemo, useState } from "react";
 import { IoIosSearch } from "react-icons/io";
 import { useRouter } from "next/navigation";
-import AddToDoForm, { AddTodoFormData, AddTodoFormProps } from "./AddTodoForm";
+import { AddTodoFormData } from "./AddTodoForm";
 import { useTodos } from "../useTodos";
 import dayjs from "dayjs";
-import QuickAddTodoForm from "./QuickAddTodoForm";
 import AddTodoModal from "./AddTodoModal";
+import TodoListContainer from "./TodoListContainer";
 
 const TodoListPage = () => {
   const [selectedDate, setSelectedDate] = useState<Date>(new Date());
@@ -52,8 +50,7 @@ const TodoListPage = () => {
         events={events}
         initialCollapsed={true}
       />
-      <QuickAddTodoForm onSubmit={handleAddTodoSubmit} />
-      <TodoList todos={todos ?? []} selectedDate={selectedDate} />
+      <TodoListContainer todos={todos ?? []} selectedDate={selectedDate} onSubmit={handleAddTodoSubmit} />
       <AddTodoModal onSubmit={handleAddTodoSubmit} selectedDate={selectedDate} />
     </div>
   );
